+++ package.json
@@ -1,36 +1,45 @@
 {
-	"name": "42_somoim",
-	"version": "1.0.0",
-	"description": "팀명: 미안하다. 이거 보여주려고 어그로 끌었다.\r     프로젝트명: 42 소모임 슬랙봇",
-	"main": "index.js",
-	"dependencies": {
-		"@slack/bolt": "^2.0.1",
-		"dotenv": "^8.2.0",
-		"eslint-config-prettier": "^6.11.0",
-		"eslint-plugin-prettier": "^3.1.3",
-		"mysql2": "^2.1.0",
-		"nodemon": "^2.0.3",
-		"prettier": "^2.0.5",
-		"sequelize": "^5.21.7",
-		"typescript": "^3.8.3",
-		"pg": "^8.0.3"
-	},
-	"devDependencies": {
-		"eslint": "^6.8.0",
-		"eslint-config-airbnb-base": "^14.1.0",
-		"eslint-plugin-import": "^2.20.2"
-	},
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/jho2301/42_Somoim.git"
-	},
-	"author": "",
-	"license": "ISC",
-	"bugs": {
-		"url": "https://github.com/jho2301/42_Somoim/issues"
-	},
-	"homepage": "https://github.com/jho2301/42_Somoim#readme"
+  "name": "42_somoim",
+  "version": "1.0.0",
+  "description": "팀명: 미안하다. 이거 보여주려고 어그로 끌었다.\r     프로젝트명: 42 소모임 슬랙봇",
+  "main": "index.js",
+  "dependencies": {
+    "@slack/bolt": "^2.0.1",
+    "@slack/web-api": "^5.8.1",
+    "@types/bluebird": "^3.5.30",
+    "@types/dotenv": "^8.2.0",
+    "@types/node": "^13.13.5",
+    "@types/validator": "^13.0.0",
+    "dotenv": "^8.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "mysql2": "^2.1.0",
+    "nodemon": "^2.0.3",
+    "pg": "^8.0.3",
+    "prettier": "^2.0.5",
+    "reflect-metadata": "^0.1.13",
+    "sequelize": "^5.21.7",
+    "sequelize-typescript": "^1.1.0",
+    "typescript": "^3.8.3"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^2.31.0",
+    "@typescript-eslint/parser": "^2.31.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.1.0",
+    "eslint-plugin-import": "^2.20.2"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jho2301/42_Somoim.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/jho2301/42_Somoim/issues"
+  },
+  "homepage": "https://github.com/jho2301/42_Somoim#readme"
 }
