+++ src/interface/index.ts
@@ -1,0 +1,9 @@
+import { WebAPICallResult } from '@slack/web-api';
+
+export interface UsersInfoResult extends WebAPICallResult {
+  user?: {
+    profile: {
+      email?: string;
+    };
+  };
+}
