--- src/skills/index.js
@@ -1,9 +1,0 @@
-const { showRegisterModal } = require('./register');
-const { showUnregisterModal } = require('./unregister');
-const { showListMessage } = require('./show_list');
-const { showHelpMessage } = require('./help');
-
-exports.showHelpMessage = showHelpMessage;
-exports.showUnregisterModal = showUnregisterModal;
-exports.showListMessage = showListMessage;
-exports.showRegisterModal = showRegisterModal;
